Place all routes into a controller folder for better organization
